requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[argon2]
python-jose[cryptography]
python-dotenv
